pyproject.toml
[project]
name = "uv-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-deepseek>=1.0.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
